requirements.txt
python-dotenv
sqlparse
groq
argparse
streamlit